+++ ewoksserver/resources/json/utils.py
@@ -3,9 +3,10 @@ import logging
 from pathlib import Path
 from typing import Iterable, Union
 
+
 ResourceIdentifierType = str
 ResourceUrlType = Path
-ResourceType = dict
+ResourceContentType = dict
 
 _logger = logging.getLogger(__name__)
 
@@ -26,7 +27,7 @@ def resource_identifiers(root: ResourceUrlType) -> Iterable[ResourceIdentifierTy
             yield _url_to_identifier(child)
 
 
-def resources(root: ResourceUrlType) -> Iterable[ResourceType]:
+def resources(root: ResourceUrlType) -> Iterable[ResourceContentType]:
     if not root.exists():
         return
     for child in root.iterdir():
@@ -39,7 +40,9 @@ def resource_exists(root: ResourceUrlType, identifier: ResourceIdentifierType) -
 
 
 def save_resource(
-    root: ResourceUrlType, identifier: ResourceIdentifierType, resource: ResourceType
+    root: ResourceUrlType,
+    identifier: ResourceIdentifierType,
+    resource: ResourceContentType,
 ):
     url = _identifier_to_url(root, identifier)
     _save_url(url, resource)
@@ -47,7 +50,7 @@ def save_resource(
 
 def load_resource(
     root: ResourceUrlType, identifier: ResourceIdentifierType
-) -> ResourceType:
+) -> ResourceContentType:
     url = _identifier_to_url(root, identifier)
     return _load_url(url)
 
@@ -65,14 +68,14 @@ def _url_to_identifier(url: ResourceUrlType) -> ResourceIdentifierType:
     return url.stem
 
 
-def _save_url(url: ResourceUrlType, resource: ResourceType):
+def _save_url(url: ResourceUrlType, resource: ResourceContentType):
     _logger.debug("Save file '%s'", url)
     url.parent.mkdir(parents=True, exist_ok=True)
     with open(url, "w") as f:
         json.dump(resource, f, indent=2)
 
 
-def _load_url(url: ResourceUrlType) -> ResourceType:
+def _load_url(url: ResourceUrlType) -> ResourceContentType:
     try:
         with open(url, "r") as f:
             return json.load(f)
@@ -81,7 +84,7 @@ def _load_url(url: ResourceUrlType) -> ResourceType:
         raise
 
 
-def _delete_url(url: ResourceUrlType) -> ResourceType:
+def _delete_url(url: ResourceUrlType) -> ResourceContentType:
     if url.exists():
         _logger.debug("Delete file '%s'", url)
         url.unlink()