+++ ewoksserver/tests/test_tasks.py
@@ -4,11 +4,20 @@ def test_single_task(serverside_client):
     response = serverside_client.get(f"/task/{identifier}")
     assert response.status_code == 404
 
-    task1a = {"task_identifier": identifier, "task_type": "class", "input_names": ["a"]}
-    response = serverside_client.put(f"/task/{identifier}", json=task1a)
+    task1a = {
+        "task_identifier": identifier,
+        "task_type": "class",
+        "required_input_names": ["a"],
+    }
+    response = serverside_client.post("/tasks", json=task1a)
     data = response.get_json()
     assert response.status_code == 200, data
-    assert data == identifier
+    expected = {
+        "required_input_names": ["a"],
+        "task_identifier": "myproject.tasks.Dummy",
+        "task_type": "class",
+    }
+    assert data == expected
 
     response = serverside_client.get(f"/task/{identifier}")
     data = response.get_json()
@@ -18,12 +27,17 @@ def test_single_task(serverside_client):
     task1b = {
         "task_identifier": identifier,
         "task_type": "class",
-        "input_names": ["a", "b"],
+        "required_input_names": ["a", "b"],
     }
     response = serverside_client.put(f"/task/{identifier}", json=task1b)
     data = response.get_json()
     assert response.status_code == 200, data
-    assert data == identifier
+    expected = {
+        "required_input_names": ["a", "b"],
+        "task_identifier": "myproject.tasks.Dummy",
+        "task_type": "class",
+    }
+    assert data == expected
 
     response = serverside_client.get(f"/task/{identifier}")
     data = response.get_json()
@@ -33,91 +47,110 @@ def test_single_task(serverside_client):
     response = serverside_client.delete(f"/task/{identifier}")
     data = response.get_json()
     assert response.status_code == 200
-    assert data == identifier
+    assert data == {"identifier": identifier}
 
     response = serverside_client.delete(f"/task/{identifier}")
     data = response.get_json()
     assert response.status_code == 200
-    assert data == identifier
+    assert data == {"identifier": identifier}
 
     response = serverside_client.get(f"/task/{identifier}")
     data = response.get_json()
     assert response.status_code == 404
-    assert data == f"task '{identifier}' does not exist"
+    expected = {
+        "identifier": identifier,
+        "message": f"Task '{identifier}' is not found.",
+        "type": "task",
+    }
+    assert data == expected
 
 
 def test_multiple_tasks(serverside_client):
     response = serverside_client.get("/tasks")
     data = response.get_json()
     assert response.status_code == 200
-    assert data == []
+    assert data == {"identifiers": []}
 
     task1a = {
         "task_identifier": "myproject.tasks.Dummy1",
         "task_type": "class",
-        "input_names": ["a"],
+        "required_input_names": ["a"],
     }
     task1b = {
         "task_identifier": "myproject.tasks.Dummy1",
         "task_type": "class",
-        "input_names": ["a", "b"],
+        "required_input_names": ["a", "b"],
     }
     task2 = {
         "task_identifier": "myproject.tasks.Dummy2",
         "task_type": "class",
-        "input_names": ["a", "b"],
+        "required_input_names": ["a", "b"],
     }
 
     response = serverside_client.post("/tasks", json=task1a)
     data = response.get_json()
     assert response.status_code == 200, data
+    assert data == task1a
+
     response = serverside_client.post("/tasks", json=task1b)
     data = response.get_json()
-    assert response.status_code == 403, data
-    assert (
-        data
-        == "Task 'myproject.tasks.Dummy1' exists. Please change identifier and retry."
-    )
+    assert response.status_code == 409, data
+    expected = {
+        "identifier": "myproject.tasks.Dummy1",
+        "message": "Task 'myproject.tasks.Dummy1' already exists.",
+        "type": "task",
+    }
+    assert data == expected
+
     response = serverside_client.post("/tasks", json=task2)
     data = response.get_json()
     assert response.status_code == 200, data
+    assert data == task2
 
     response = serverside_client.get("/tasks")
     data = response.get_json()
     assert response.status_code == 200, data
     expected = {"myproject.tasks.Dummy1", "myproject.tasks.Dummy2"}
-    assert set(data) == expected
+    assert set(data["identifiers"]) == expected
 
 
 def test_discover_tasks(serverside_client):
     response = serverside_client.get("/tasks")
     data = response.get_json()
     assert response.status_code == 200
-    assert data == []
+    assert data == {"identifiers": []}
 
     module = "ewoksserver.tests.dummy_tasks"
 
     response = serverside_client.post("/tasks/discover", json={"module": module})
     data = response.get_json()
     assert response.status_code == 200, data
+    expected = {
+        "ewoksserver.tests.dummy_tasks.MyTask1",
+        "ewoksserver.tests.dummy_tasks.MyTask2",
+    }
+    assert set(data["identifiers"]) == expected
 
     response = serverside_client.get("/tasks")
-    identifiers = response.get_json()
+    data = response.get_json()
     assert response.status_code == 200
-    assert identifiers
-    assert all(identifier.startswith(module) for identifier in identifiers)
+    expected = {
+        "ewoksserver.tests.dummy_tasks.MyTask1",
+        "ewoksserver.tests.dummy_tasks.MyTask2",
+    }
+    assert set(data["identifiers"]) == expected
 
     response = serverside_client.post("/tasks/discover", json={"module": module})
     data = response.get_json()
-    assert response.status_code == 403, data
-    assert "Please change identifier and retry." in data
+    assert response.status_code == 409, data
+    assert "already exists" in data["message"]
 
 
 def test_task_descriptions(serverside_client):
     response = serverside_client.get("/tasks/descriptions")
     data = response.get_json()
     assert response.status_code == 200
-    assert data == []
+    assert data == {"items": []}
 
     module = "ewoksserver.tests.dummy_tasks"
 
@@ -126,9 +159,9 @@ def test_task_descriptions(serverside_client):
     assert response.status_code == 200, data1
 
     response = serverside_client.get("/tasks/descriptions")
-    data2 = response.get_json()
+    data2 = response.get_json()["items"]
     data2 = {
         r["task_identifier"] for r in data2 if r["task_identifier"].startswith(module)
     }
     assert response.status_code == 200
-    assert set(data1) == data2
+    assert set(data1["identifiers"]) == data2