+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # CHANGELOG.md
 
-## 0.0.1 (unreleased)
+## 0.1.0 (unreleased)
 
 Added:
-  - 
+  - REST API to load and store ewoks workflows
+  - REST API to load and store ewoks tasks
+  - REST API to Execute workflows with ewoksjob
+  - Websocket API to ewoks execution events
+  - Serve ewoksweb frontend